# chacha20: Remove mutable borrows from AVX2 backend