@str4d 2021-08-29
When compiled with `RUSTFLAGS="-Ctarget-feature=+avx2"`, this PR generates almost exactly the same assembly as current master (a handful operations are reordered), so the changes only affect autodetect mode (and apparently not materially on my machine, but maybe it's useful on others).